bills ok --ingreso de licors by day

+++ db/migrate/20160225231036_create_licors.rb
@@ -2,7 +2,7 @@ class CreateLicors < ActiveRecord::Migration
   def change
     create_table :licors do |t|
       t.string :nombre
-      t.float :mililitros
+      t.integer :mililitros
       t.float :valor
       t.belongs_to :category_licor, index: true, foreign_key: true
       t.text :descripcion
